+++ R/query_db.R
@@ -15,63 +15,27 @@
 #' @export
 #'
 #' @examples
-#' conn <- connect_to_db(":memory:")
+#' \dontrun{
+#' conn <- connect_to_db("path/to/database.db")
 #'
-#' mtcars$mtcars_id = 1:nrow(mtcars)
+#' arguments <- add_argument(list(), "publication_id", "greater", 0)
 #'
-#' example_data = data.frame(
-#'   example_id = 1:150,
-#'   mtcars_id = rep(1:30, each = 5),
-#'   example_value = runif(150, 0, 1)
-#' )
+#' target_vars <- c("default", "n_participants")
 #'
-#' DBI::dbWriteTable(conn, "mtcars_table", mtcars)
-#' DBI::dbWriteTable(conn, "example_table", example_data)
+#' target_table <- "study_table"
 #'
-#' # Initializing argument list
-#' arguments = list()
-#' arguments = add_argument(
-#'  list = arguments,
-#'  conn = conn,
-#'  variable = "cyl",
-#'  operator = "equal",
-#'  values = c(4, 6)
-#' )
-#'
-#' arguments = add_argument(
-#'  list = arguments,
-#'  conn = conn,
-#'  variable = "example_value",
-#'  operator = "greater",
-#'  values = 0.4
-#' )
-#'
-#' # Return specified variables
-#' target_vars = c("mtcars_id", "example_id", "cyl")
-#'
-#' query_results = query_db(
-#'  conn = conn,
-#'  arguments = arguments,
-#'  target_vars = target_vars,
-#'  target_table = "example_table",
-#'  argument_relation = "and"
-#' )
-#'
-#' # Return all variables in mtcars_table and example_value from example_table
-#' query_results = query_db(
-#'  conn = conn,
-#'  arguments = arguments,
-#'  target_vars = c("default", "example_value"),
-#'  target_table = "mtcars_table",
-#'  argument_relation = "and"
-#' )
-#'
+#' results <- query_db(conn, arguments, target_vars, target_table)
+#' }
 query_db <- function(conn, arguments, target_vars = "default", target_table = "observation_table", argument_relation = "and"){
   # Convert argument_relation into proper numerical vector
   argument_sequence = get_argument_sequence(arguments, argument_relation)
 
   # Get info about the db structure
   col_names = get_column_names(conn)
+
+  if (!(target_table %in% col_names$table)){
+    stop(paste("Target table", target_table, "not found in database. Available tables:", paste(unique(col_names$table), collapse = ", ")))
+  }
 
   # replace "default" in target_vars with all variables of the target table
   target_table_vars = col_names[col_names$table == target_table, "column"]